CursosDjango_OasisNegro/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==12.1.1
PyMuPDF==1.27.2
PyPDF2==3.0.1
sqlparse==0.5.5
tzdata==2025.3
